migrate fields to pysdk
